requirements.txt
aws-cdk-lib==2.190.0
constructs>=10.0.0,<11.0.0
constructs>=10.0.0,<11.0.0
omegaconf==2.3.0
boto3~=1.38.11
botocore~=1.38.11
